requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.5
openai==1.55.0
python-dotenv==1.0.1
uvicorn==0.32.1
httpx==0.27.2